+++ plugins/content-manager/src/main/java/com/wazuh/contentmanager/transport/TransportUpdateIntegrationAction.java
@@ -37,6 +37,7 @@ import com.wazuh.contentmanager.action.UpdateIntegrationAction;
 import com.wazuh.contentmanager.cti.catalog.index.ContentIndex;
 import com.wazuh.contentmanager.cti.catalog.model.Resource;
 import com.wazuh.contentmanager.cti.catalog.model.Space;
+import com.wazuh.contentmanager.cti.catalog.service.EngineContentLoader;
 import com.wazuh.contentmanager.cti.catalog.service.SecurityAnalyticsService;
 import com.wazuh.contentmanager.cti.catalog.service.SecurityAnalyticsServiceImpl;
 import com.wazuh.contentmanager.engine.service.EngineService;
@@ -71,8 +72,15 @@ public class TransportUpdateIntegrationAction extends AbstractTransportUpdateAct
             TransportService transportService,
             ActionFilters actionFilters,
             Client client,
-            EngineService engine) {
-        super(UpdateIntegrationAction.NAME, transportService, actionFilters, client, engine);
+            EngineService engine,
+            EngineContentLoader engineContentLoader) {
+        super(
+                UpdateIntegrationAction.NAME,
+                transportService,
+                actionFilters,
+                client,
+                engine,
+                engineContentLoader);
     }
 
     @Override