+++ CHANGELOG.md
@@ -64,6 +64,8 @@
 - Safeguard sensitive configurations [(#1277)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1277)
 - Reduce TTL for historical information about deletes in-memory [(#1328)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1328)
 - Integration's mode [(#1356)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1356)
+- Load standard space into all cluster nodes [(#1376)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1376)
+- Added `previous` value fields to the WCS for process, service, and package inventory changes [(#1413)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1413)
 
 ### Changed
 - Add dependabot scan to the content-manager plugin [(#442)](https://github.com/wazuh/wazuh-indexer-plugins/issues/442)
@@ -170,6 +172,7 @@
 - [BUG] Missing threat-intel indices [(#1362)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1362)
 - Update the Content Manager OpenAPI (`openapi.yml`) to match the current API [(#1353)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1353)
 - [BUG] Vulnerability Detection empty for all agents after a transient snapshot bootstrap failure [(#1383)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1383)
+- [BUG] Log test does not use manually enabled integrations [(#1410)](https://github.com/wazuh/wazuh-indexer-plugins/issues/1410)
 
 ## Prior versions
 - []()